% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageBinPlot.R
\name{viridisColor}
\alias{viridisColor}
\title{Helper function for getting viridis palette}
\usage{
viridisColor(
  palette = c("viridis", "inferno", "magma", "plasma", "cividis", "mako", "rocket",
    "turbo"),
  begin = 0.3,
  end = 1,
  n = 10,
  alpha = 1,
  ...
)
}
\arguments{
\item{palette}{Character string specifying the viridis palette name.}

\item{begin}{Numeric, starting hue (0-1).}

\item{end}{Numeric, ending hue (0-1).}

\item{n}{Integer, number of colors.}

\item{alpha}{Numeric, alpha transparency (0-1).}

\item{...}{Other arguments passed to the palette function.}
}
\value{
A character vector of colors.
}
\description{
Helper function for getting viridis palette
}
