% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageBinPlot.R
\name{viridisColor}
\alias{viridisColor}
\title{viridisColor}
\usage{
viridisColor(
  palette = c("viridis", "inferno", "magma", "plasma", "cividis", "mako", "rocket",
    "turbo"),
  begin = 0.3,
  end = 1,
  n = 10,
  alpha = 1,
  ...
)
}
\arguments{
\item{palette}{Character, one of 'viridis', 'inferno', 'magma', 'plasma', 'cividis', 'mako', 'rocket', 'turbo'.}

\item{begin}{Numeric, starting hue (between 0 and 1). Default is 0.3.}

\item{end}{Numeric, ending hue (between 0 and 1). Default is 1.}

\item{n}{Integer, number of colors to generate. Default is 10.}

\item{alpha}{Numeric, transparency level (between 0 and 1). Default is 1.}

\item{...}{Additional arguments passed to the palette function.}
}
\value{
A vector of colors from the specified viridis palette.
}
\description{
Helper function for getting viridis palette
}
\keyword{internal}
